+++ src/main/resources/l10n/JabRef_de.properties
@@ -2430,4 +2430,5 @@ Executing_command_\"%0\"...=Ausführung_des_Kommandos_\"%0\"...
 Error_occured_while_executing_the_command_\"%0\".=Während_der_Ausführung_des_Befehls_\"%0\"_ist_ein_Fehler_aufgetreten.
 
 
-Add_possibly_missing_dash_to_ISSN=
+
+Reformat_ISSN=
