+++ src/main/resources/l10n/JabRef_ja.properties
@@ -2379,3 +2379,5 @@ Note\:_Use_the_placeholder_%0_for_the_location_of_the_opened_database_file.=
 Executing_command_\"%0\"...=
 Error_occured_while_executing_the_command_\"%0\".=
 Add_possibly_missing_dash_to_ISSN=
+
+Reformat_ISSN=
